+++ src/main/java/io/github/torand/jsonschema2java/collectors/TypeInfoCollector.java
@@ -28,6 +28,7 @@ import java.util.Optional;
 import static io.github.torand.javacommons.collection.CollectionHelper.isEmpty;
 import static io.github.torand.javacommons.lang.Exceptions.illegalStateException;
 import static io.github.torand.javacommons.lang.StringHelper.nonBlank;
+import static io.github.torand.jsonschema2java.collectors.Extensions.EXT_JSON_DESERIALIZER;
 import static io.github.torand.jsonschema2java.collectors.Extensions.EXT_JSON_FORMAT;
 import static io.github.torand.jsonschema2java.collectors.Extensions.EXT_JSON_SERIALIZER;
 import static io.github.torand.jsonschema2java.collectors.Extensions.EXT_NULLABLE;
@@ -156,6 +157,12 @@ public class TypeInfoCollector extends BaseCollector {
         if (maybeJsonSerializer.isPresent()) {
             AnnotationInfo jsonSerializeAnnotation = getJsonSerializeAnnotation(maybeJsonSerializer.get());
             typeInfo = typeInfo.withAddedAnnotation(jsonSerializeAnnotation);
+        }
+
+        Optional<String> maybeJsonDeserializer = schema.extensions().getString(EXT_JSON_DESERIALIZER);
+        if (maybeJsonDeserializer.isPresent()) {
+            AnnotationInfo jsonDeserializeAnnotation = getJsonDeserializeAnnotation(maybeJsonDeserializer.get());
+            typeInfo = typeInfo.withAddedAnnotation(jsonDeserializeAnnotation);
         }
 
         Optional<String> maybeValidationConstraint = schema.extensions().getString(EXT_VALIDATION_CONSTRAINT);
@@ -421,9 +428,15 @@ public class TypeInfoCollector extends BaseCollector {
     }
 
     private AnnotationInfo getJsonSerializeAnnotation(String jsonSerializer) {
-        return new AnnotationInfo("@JsonSerialize(using = %s)".formatted(getJsonSerializerClass(jsonSerializer)))
+        return new AnnotationInfo("@JsonSerialize(using = %s)".formatted(getClassRefFromFullyQualifiedClassName(jsonSerializer)))
             .withAddedNormalImport("com.fasterxml.jackson.databind.annotation.JsonSerialize")
             .withAddedNormalImport(jsonSerializer);
+    }
+
+    private AnnotationInfo getJsonDeserializeAnnotation(String jsonDeserializer) {
+        return new AnnotationInfo("@JsonDeserialize(using = %s)".formatted(getClassRefFromFullyQualifiedClassName(jsonDeserializer)))
+            .withAddedNormalImport("com.fasterxml.jackson.databind.annotation.JsonDeserialize")
+            .withAddedNormalImport(jsonDeserializer);
     }
 
     private AnnotationInfo getJsonFormatAnnotation(String pattern) {
@@ -506,8 +519,8 @@ public class TypeInfoCollector extends BaseCollector {
             "jakarta.validation.constraints.Size");
     }
 
-    private String getJsonSerializerClass(String jsonSerializerFqn) {
-        String className = getClassNameFromFqn(jsonSerializerFqn);
+    private String getClassRefFromFullyQualifiedClassName(String fullyQualifiedClassName) {
+        String className = getClassNameFromFqn(fullyQualifiedClassName);
         return formatClassRef(className);
     }
 }